deploy: e88ce8af4f62259d2798a6dc35f691d7e4cd6f2c

+++ RSGL.h
@@ -108,8 +108,8 @@ RSGL basicDraw types
 		typedef signed short 	i16;
 		typedef unsigned int 	u32;
 		typedef signed int		i32;
-		typedef unsigned long	u64;
-		typedef signed long		i64;
+		typedef unsigned long long	u64;
+		typedef signed   long long	i64;
 	#else
 		#include <stdint.h>
